+++ Sources/CommandLineWrapper/BuildEngineCommand.swift
@@ -1,10 +1,3 @@
-//
-//  BuildEngineCommand.swift
-//  ContentBlockerConverter
-//
-//  Created by Andrey Meshkov on 12/04/2025.
-//
-
 import ArgumentParser
 import ContentBlockerConverter
 import FilterEngine
@@ -56,23 +49,12 @@ struct BuildEngineCommand: ParsableCommand {
         Logger.log("(BuildEngine) - Advanced rules count: \(rules.count)")
 
         // Build the FilterEngine binary using the provided rules and safari version.
-        // Remove any previous settings from sharedUserDefaults.
-        guard let sharedUserDefaults = EmptyDefaults(suiteName: #file) else {
-            throw NSError(
-                domain: "UserDefaultsError",
-                code: 1,
-                userInfo: [NSLocalizedDescriptionKey: "Failed to create UserDefaults"]
-            )
-        }
-        defer { sharedUserDefaults.removePersistentDomain(forName: #file) }
-
         // Create a URL from the output directory string.
         let containerURL = URL(fileURLWithPath: outputDir)
 
         // Pass the URL to the WebExtension constructor.
         let webExtension = try WebExtension(
             containerURL: containerURL,
-            sharedUserDefaults: sharedUserDefaults,
             version: safariVersionResolved
         )
 
@@ -80,23 +62,5 @@ struct BuildEngineCommand: ParsableCommand {
         _ = try webExtension.buildFilterEngine(rules: rules.joined(separator: "\n"))
 
         Logger.log("(BuildEngine) - FilterEngine files written to \(outputDir)")
-    }
-}
-
-/// EmptyDefaults is a UserDefaults instance that does not save or read anything.
-/// It is required to avoid creating unnecessary files.
-class EmptyDefaults: UserDefaults {
-    override func double(forKey defaultName: String) -> Double {
-        0
-    }
-
-    override func integer(forKey defaultName: String) -> Int {
-        0
-    }
-
-    override func set(_ value: Double, forKey defaultName: String) {
-    }
-
-    override func set(_ value: Int, forKey defaultName: String) {
     }
 }