+++ CHANGELOG.md
@@ -7,7 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-Nothing here so far.
+### Changed
+
+- `WebExtension` does not require shared `UserDefaults` anymore: [#91]
+- `WebExtension.shared()` now uses both group and version as a key: [#84]
+- Web extension schema migration now takes into account that the extension
+  process can be interrupted and in this case it won't make another attempt
+  to migrate.
+- Updated scriptlets dependency to `2.2.8`: [#92]
+
+[#84]: https://github.com/AdguardTeam/SafariConverterLib/issues/84
+[#91]: https://github.com/AdguardTeam/SafariConverterLib/issues/91
+[#92]: https://github.com/AdguardTeam/SafariConverterLib/issues/92
+
+### Fixed
+
+- Fixed bugs with parsing scriptlet arguments: 1-character arguments, escaped quotes: [#93]
+
+[#93]: https://github.com/AdguardTeam/SafariConverterLib/issues/93
 
 ## [v3.0.1]